+++ packages/react/src/datepicker/Datepicker.tsx
@@ -1,6 +1,6 @@
 import {
   Input as KVInput,
-  InputProps as KVInputProps,
+  InputProps,
   Popover,
   useBoolean,
   PopoverContent,
@@ -8,17 +8,17 @@ import {
   InputRightElement,
   PopoverTrigger,
   PopoverAnchor,
+  theme,
+  IconButton,
 } from "@kvib/react/src";
-import { forwardRef } from "@chakra-ui/react";
+import { forwardRef, useFormControlContext } from "@chakra-ui/react";
 import { DayPicker, useInput } from "react-day-picker";
 import "react-day-picker/dist/style.css";
-import "./Datepicker.css";
-import { Icon } from "@kvib/react/src/icon";
 import nb from "date-fns/locale/nb/index.js";
 import { ChangeEvent, useEffect } from "react";
 import { isValid } from "date-fns";
 
-type DatepickerProps = KVInputProps & {
+export type DatepickerProps = Omit<InputProps, "colorScheme" | "max" | "min" | "defaultValue"> & {
   /**
    * A default date to be selected when the picker is displayed.
    */
@@ -73,43 +73,58 @@ type DatepickerProps = KVInputProps & {
    * Sideeffect to be run when a date is selected.
    */
   onChange?: (date: Date | undefined) => void;
+
+  /**
+   * Whether or not the input is disabled.
+   */
+  isDisabled?: boolean;
+
+  /**
+   * Whether or not the input is invalid.
+   */
+  isInvalid?: boolean;
+
+  /**
+   * Whether or not the input is required.
+   */
+  isRequired?: boolean;
+
+  /**
+   * The colorScheme for the Datepicker.
+   */
+  colorScheme?: "blue" | "green";
 };
 
-type ExcludedProps = "max" | "min" | "defaultValue";
+export const Datepicker = forwardRef<DatepickerProps, "input">(({ onChange, useNative = true, ...props }, ref) => {
+  const KVInputProps = extractKVProps(props);
+  const commonProps = getCommonInputProps(props);
+  const defaultValue = props.defaultSelected ? formatDate(props.defaultSelected) : undefined;
+  const isClient = typeof window === "object";
+  const isMobile = isClient ? window.innerWidth < 480 : false;
 
-type DatepickerPropsWithoutStandard = Omit<DatepickerProps, ExcludedProps>;
+  const handleNativeChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const date = new Date(event.target.value);
+    if (isValid(date)) {
+      onChange?.(date);
+    }
+  };
 
-export const Datepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
-  ({ onChange, useNative = true, ...props }, ref) => {
-    const KVInputProps = extractKVProps(props);
-    const commonProps = getCommonInputProps(props);
-    const defaultValue = props.defaultSelected ? formatDate(props.defaultSelected) : undefined;
-    const isClient = typeof window === "object";
-    const isMobile = isClient ? window.innerWidth < 480 : false;
+  if (isMobile && useNative)
+    return (
+      <KVInput
+        ref={ref}
+        type="date"
+        defaultValue={defaultValue}
+        {...KVInputProps}
+        {...commonProps}
+        onChange={handleNativeChange}
+      />
+    );
 
-    const handleNativeChange = (event: ChangeEvent<HTMLInputElement>) => {
-      const date = new Date(event.target.value);
-      if (isValid(date)) {
-        onChange?.(date);
-      }
-    };
+  return <CustomDatepicker {...props} {...commonProps} onChange={onChange} />;
+});
 
-    if (isMobile && useNative)
-      return (
-        <KVInput
-          type="date"
-          defaultValue={defaultValue}
-          {...KVInputProps}
-          {...commonProps}
-          onChange={handleNativeChange}
-        />
-      );
-
-    return <CustomDatepicker {...props} ref={ref} {...commonProps} onChange={onChange} />;
-  },
-);
-
-const CustomDatepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
+const CustomDatepicker = forwardRef<DatepickerProps, "input">(
   (
     {
       onChange,
@@ -122,10 +137,30 @@ const CustomDatepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
       showOutsideDays,
       showWeekNumber,
       disabledDays,
+      isDisabled: isDisabledExternally = false,
+      isInvalid: isInvalidExternally = false,
+      isRequired: isRequiredExternally = false,
+      colorScheme = "green",
       ...KVInputProps
     },
     ref,
   ) => {
+    // Style for the day picker
+    const uniqueClassName = generateUniqueClassName("kvib-datepicker");
+    const style = css(uniqueClassName, colorScheme);
+
+    // Get state from form control context
+    const formControlContext = useFormControlContext();
+    const isDisabledFromForm = formControlContext?.isDisabled || false;
+    const isInvalidFromForm = formControlContext?.isInvalid || false;
+    const isRequiredFromForm = formControlContext?.isRequired || false;
+
+    // Determine the effective isDisabled, isInvalid and isRequired states
+    const isDisabled = isDisabledExternally || isDisabledFromForm;
+    const isInvalid = isInvalidExternally || isInvalidFromForm;
+    const isRequired = isRequiredExternally || isRequiredFromForm;
+
+    // State for the day picker
     const [isPickerVisible, setPickerVisible] = useBoolean(false);
     const { inputProps, dayPickerProps } = useInput({
       defaultSelected,
@@ -133,12 +168,14 @@ const CustomDatepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
       locale: nb,
       fromDate,
       toDate,
+      required: isRequired,
     });
 
     useEffect(() => {
       // Check if the selected date in the day picker has changed
       if (dayPickerProps.selected) {
         onChange?.(dayPickerProps.selected);
+        setPickerVisible.off();
       }
     }, [dayPickerProps.selected, onChange]);
 
@@ -151,26 +188,31 @@ const CustomDatepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
       >
         <InputGroup>
           <PopoverAnchor>
-            <KVInput ref={ref} className="custom-datepicker" {...KVInputProps} {...inputProps} />
+            <KVInput
+              ref={ref}
+              className="custom-datepicker"
+              isDisabled={isDisabled}
+              isInvalid={isInvalid}
+              isRequired={isRequired}
+              {...KVInputProps}
+              {...inputProps}
+            />
           </PopoverAnchor>
-          <InputRightElement height="100%">
+          <InputRightElement opacity={isDisabled ? 0.5 : 1} pointerEvents={isDisabled ? "none" : "auto"} height="100%">
             <PopoverTrigger>
-              <button
-                style={{
-                  height: "90%",
-                  width: "100%",
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                }}
+              <IconButton
+                icon="calendar_today"
+                colorScheme={colorScheme as "blue" | "green"}
+                size={KVInputProps.size}
+                aria-label="open datepicker"
                 onClick={setPickerVisible.toggle}
-              >
-                <Icon icon="calendar_today" size={KVInputProps.size === "xs" || KVInputProps.size === "sm" ? 20 : 24} />
-              </button>
+                variant="tertiary"
+              />
             </PopoverTrigger>
           </InputRightElement>
         </InputGroup>
-        <PopoverContent width="auto">
+        <PopoverContent width="auto" padding="1rem">
+          <style>{style}</style>
           <DayPicker
             captionLayout={showDropdownMonthYear ? "dropdown-buttons" : undefined}
             disableNavigation={disableNavigation}
@@ -178,6 +220,7 @@ const CustomDatepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
             showOutsideDays={showOutsideDays}
             showWeekNumber={showWeekNumber}
             disabled={disabledDays}
+            classNames={{ root: uniqueClassName }}
             {...dayPickerProps}
           />
         </PopoverContent>
@@ -186,7 +229,8 @@ const CustomDatepicker = forwardRef<DatepickerPropsWithoutStandard, "input">(
   },
 );
 
-function extractKVProps(props: DatepickerProps): KVInputProps {
+// Function to extract the props that are used by the KVInput (native) component
+function extractKVProps(props: DatepickerProps): InputProps {
   const {
     defaultSelected,
     defaultMonth,
@@ -205,6 +249,7 @@ function extractKVProps(props: DatepickerProps): KVInputProps {
 
 type ValidDateInput = number | Date | string;
 
+// Function to format a date to the format used by the datepicker
 function formatDate(date: ValidDateInput): string {
   let dateObject: Date;
 
@@ -227,6 +272,7 @@ function formatDate(date: ValidDateInput): string {
   return `${y}-${m}-${d}`;
 }
 
+// Function to get the common input props the native and custom datepicker
 const getCommonInputProps = (props: DatepickerProps) => {
   const min = props.fromDate ? formatDate(props.fromDate) : undefined;
   const max = props.toDate ? formatDate(props.toDate) : undefined;
@@ -235,4 +281,24 @@ const getCommonInputProps = (props: DatepickerProps) => {
     min,
     max,
   };
+};
+
+// Function to generate the css for the day picker
+const css = (className: string, colorScheme: "blue" | "green") => {
+  return `
+ .${className} {
+  --rdp-cell-size: 40px; /* Size of the day cells. */
+  --rdp-caption-font-size: 18px; /* Font size for the caption labels. */
+  --rdp-accent-color: ${theme.colors[colorScheme][500]}; /* Accent color for the background of selected days. */
+  --rdp-background-color: ${theme.colors[colorScheme][50]}; /* Background color for the hovered/focused elements. */
+  --rdp-outline: 2px solid var(--rdp-accent-color); /* Outline border for focused elements */
+  --rdp-selected-color: #fff; /* Color of selected day text */
+}
+`;
+};
+
+// Function to generate a unique class name
+const generateUniqueClassName = (baseName: string) => {
+  const uniquePart = Math.random().toString(36).substring(2, 7); // Generates a random string
+  return `${baseName}-${uniquePart}`;
 };